+++ CHANGELOG.md
@@ -5,11 +5,15 @@
 ### API changes
 
 - Add "mode" option for crud select operation ([#107](https://github.com/tarantool/cartridge-java/issues/107))
-- Change using of proxy client parameters (mode, rollback_on_error, stop_on_error) with enum classes([#419](https://github.com/tarantool/cartridge-java/issues/419))
+- Change using of proxy client parameters (mode, rollback_on_error, stop_on_error) with enum classes ([#419](https://github.com/tarantool/cartridge-java/issues/419))
 
 ### Bugfixes
 
 - Fix Instant converter to parse 8 bytes datetime ([#408](https://github.com/tarantool/cartridge-java/issues/408))
+
+### Internal changes
+
+- Change using of option names with the String type in the internal API to using with the ProxyOption enum class ([#420](https://github.com/tarantool/cartridge-java/issues/420))
 
 ## [0.12.1] - 2023-08-04